+++ server/api/availability/helper-functions.ts
@@ -1,6 +1,6 @@
 
 import { SplitDate } from "../../../shared/types/splitDate"
-import { AvailabilityWeekWithAvailibilityWithShiftTypes, AvailabilityWithShiftTypes } from "../../utils/schedule-generator-types/availibility"
+import { AvailabilityWeekWithAvailibilityWithShiftTypes, AvailabilityWithShiftTypes } from "../../types/availibility"
 
 export const getAvailabilityForDate = (availabilityWeeks: AvailabilityWeekWithAvailibilityWithShiftTypes[], date: SplitDate): AvailabilityWithShiftTypes | undefined => {
   if (!availabilityWeeks || availabilityWeeks.length <= 0) return undefined
