+++ server/api/availability/router.ts
