--- server/api/required-staffing/index.ts
@@ -1,1 +1,0 @@
-export { requiredStaffingRouter } from "./required-staffing-router"
