+++ server/api/root.ts
@@ -1,14 +1,14 @@
 import { createTRPCRouter } from "./trpc"
 import { scheduleRouter } from "./schedule/"
 import { staffingRouter } from "./staffing"
-import { requiredStaffingRouter } from "./required-staffing"
+import { staffRequiredRouter } from "./staff-required"
 import { availabilityRouter } from "./availability"
 
 
 export const appRouter = createTRPCRouter({
   schedule: scheduleRouter,
   staffing: staffingRouter,
-  requiredStaffing: requiredStaffingRouter,
+  staffRequired: staffRequiredRouter,
   availability: availabilityRouter,
 })
