+++ server/api/staff-required/index.ts
@@ -1,0 +1,1 @@
+export { staffRequiredRouter } from "./router"
