+++ server/api/staffing/index.ts
@@ -1,1 +1,2 @@
-export { staffingRouter } from "./staffing-router"
+export { staffingRouter } from "./router"
+export * from "./database-actions"
