+++ server/types/availibility.ts
